feat: widen stock quantities to Numeric + unit stamp (uom step 1) (#894)
* feat: widen stock quantities to Numeric + unit stamp (uom step 1)

Step 1 of the units-of-measure track (Refs #893). A widening and nothing
else: the seven quantity columns become Numeric(18,6) in Postgres, but
every schema, annotation and frontend gate above the database still says
`int`, so the API accepts and returns whole numbers exactly as before.

Migration 0074:
  - widens stock_entries.quantity_delta, project_entries.quantity,
    order_entries.quantity_{ordered,received}, lots.purchase_quantity,
    parts.low_stock_report_quantity and parts.attrition_min_quantity;
  - adds parts.unit_of_measure and an immutable per-row stock_entries.unit
    stamp, both String(16) NOT NULL DEFAULT 'pcs'. Stamping the unit on the
    row rather than resolving it through the part on read is what stops a
    part edit retroactively reinterpreting history.

integer -> numeric is lossless, so this stays reversible until the first
fractional value is written — and nothing here can write one. downgrade()
therefore refuses on any fractional or out-of-int4-range row rather than
letting Postgres' rounding numeric -> integer cast destroy it, mirroring
the guard 0032 put on its own upgrade.

The ALTERs rewrite their tables under ACCESS EXCLUSIVE; the migration sets
a 5s lock_timeout so it fails fast instead of queueing ahead of every
reader. stock_entries needs a maintenance window.

Application side is plumbing only: quantity_out() pins the untyped route
and MCP serialisers back to JSON integers (a scaled Decimal would render
as 5.0), and builds/service.py::_required coerces the now-Decimal
attrition floor so its float arithmetic stays bit-identical.

* fix: address migration review on 0074 (uom step 1)

Five rewrites instead of seven. `ALTER COLUMN ... TYPE` reindexes the
whole table, not just the changed column, so two separate ALTERs against
`parts` rebuilt its five pg_trgm GIN indexes twice for nothing. The
widenings are now grouped into one ALTER TABLE per table, and the deploy
note says what a rewrite actually costs — indexes too, ~2x peak disk and
comparable WAL — rather than "the heap rewrite alone".

The downgrade grew two teeth:

  - it takes ACCESS EXCLUSIVE on all five tables BEFORE counting. A bare
    COUNT(*) holds only ACCESS SHARE, so a row inserted between the guard
    and the ALTER would have sailed past the check and been rounded away;
  - it refuses to drop a non-default unit stamp. A whole number is only
    safe to narrow if it still means pieces — dropping the stamp off a
    12 m spool leaves an ambiguous 12, which is the loss the per-row
    stamp exists to prevent. 0074 cannot be edited once merged, so this
    has to be here rather than in the step that opens non-'pcs' units.

Also `orders/service.py`: `outstanding` is now Decimal - Decimal and is
interpolated into an OrderError that the route re-raises verbatim as the
public 400 body, so it was about to start reading "outstanding 6.000000".

Deploy note now spells out the cron-sidecar hazard: they declare
`depends_on: backend: service_healthy`, so the OLD cron containers keep
their connections through the migration and can turn the 5s lock_timeout
into a crash-loop. Stop them for the window.

* fix: keep orders.py inside its 400-line budget

`orders.py` sat at exactly its CQ-002 cap on main, so the one import this
PR adds tipped it to 401 and failed `line-count-budget`. Reclaimed the
line from this PR's own `_totals` rather than raising the cap — the cap is
the policy, and a widening has no business spending it.

* fix: finish the chain before the ORM in test_migration_0070_backfill

Both tests walk down to 0069, back up to 0070, and then keep using the
API — but the ORM models describe `head`, not 0070. Every ORM `SELECT
parts.*` now names `parts.unit_of_measure`, which 0074 adds and 0070 does
not have, so both tests died on UndefinedColumn.

The raw-SQL assertions that actually check 0070's backfill are unmoved
and still run against 0070's schema. Only the API assertions below them
now run after `upgrade head`; nothing between 0070 and head touches
`part_provider_links` or `workspace_provider_credentials`, so what they
observe is still purely 0070's work.

Commented rather than just fixed: this is a trap for the next migration
that adds a column, not something specific to 0074.

---------

+++ backend/app/domain/lots/models.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import Column, Date, ForeignKey, Index, Integer, Numeric, String, Text
+from sqlalchemy import Column, Date, ForeignKey, Index, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID
 
 from app.domain._mixins import WorkspaceOwned
@@ -51,6 +51,10 @@ class Lot(WorkspaceOwned, Base):
         ForeignKey("builds.id", ondelete="SET NULL", name="fk_lots_source_build_id"),
         nullable=True,
     )
-    purchase_quantity = Column(Integer, nullable=True)
+    # Snapshot of the originating add. Widened to Numeric(18,6) in alembic
+    # 0074 in lockstep with `stock_entries.quantity_delta` — if the two
+    # scales diverge, a lot's purchase quantity stops agreeing with the
+    # ledger sum that produced it.
+    purchase_quantity = Column(Numeric(18, 6), nullable=True)
     purchase_unit_cost = Column(Numeric(18, 6), nullable=True)
     purchase_currency = Column(String(3), nullable=True)
